+++ archsim/inc/arch/risc-v/RiscVLinuxUserEmulationModel.h
@@ -27,7 +27,7 @@ namespace archsim
 				bool Initialise(System& system, uarch::uArch& uarch) override;
 				void Destroy() override;
 
-				archsim::abi::ExceptionAction HandleException(archsim::core::thread::ThreadInstance *cpu, unsigned int category, unsigned int data);
+				archsim::abi::ExceptionAction HandleException(archsim::core::thread::ThreadInstance *cpu, uint64_t category, uint64_t data);
 				bool InvokeSignal(int signum, uint32_t next_pc, archsim::abi::SignalData* data);
 
 				bool PrepareBoot(System& system);
